// src/interfaces.ts
export interface IGeneralInfo {
  phoneNumber: string;
  firstName: string;
  lastName: string;
  email: string;
  statement: string;
}
export interface IEducation {
  schoolName: string;
  startDate: string;
  endDate: string;
  study: string;
}
export interface IWorkHistory {
  companyName: string;
  position: string;
  startDate: string;
  endDate: string;
  description: string;
}